byokg-rag/src/requirements.txt
pydantic>=2.8.2
boto3
xmltodict
colorama
pyyaml
pytest
faiss-cpu==1.9.0
thefuzz
FlagEmbedding==1.2.11
langchain_huggingface
ipykernel
